+++ CHANGELOG.md
@@ -4,8 +4,15 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
-## [0.3.0] - 2021-10-27
+### Changed
 
+- `operator-rs` `0.3.0` → `0.4.0` ([#187]).
+- Adapted pod image and container command to docker image ([#187]).
+- Adapted documentation to represent new workflow with docker images ([#187]).
+
+[#187]: https://github.com/stackabletech/spark-operator/pull/187
+
+## [0.3.0] - 2021-10-27
 
 ### Changed
 - `operator-rs`: `0.3.0` ([#179])